% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{ggmix_data_object}
\alias{ggmix_data_object}
\alias{new_fullrank_kinship}
\alias{new_fullrank_K}
\alias{new_fullrank_UD}
\alias{new_lowrank_kinship}
\alias{new_lowrank_K}
\alias{new_lowrank_UD}
\title{Constructor functions for the different ggmix objects}
\usage{
new_fullrank_kinship(kinship)

new_fullrank_K(K)

new_fullrank_UD(U, D)

new_lowrank_kinship(kinship, n_nonzero_eigenvalues, n_zero_eigenvalues)

new_lowrank_K(K, n_nonzero_eigenvalues, n_zero_eigenvalues)

new_lowrank_UD(U, D, n_nonzero_eigenvalues, n_zero_eigenvalues)
}
\arguments{
\item{kinship}{positive definite kinship matrix}

\item{K}{the matrix of SNPs used to determine the kinship matrix}

\item{U}{left singular vectors corresponding to the eigenvalues provided in
the \code{D} argument}

\item{D}{eigenvalues}

\item{n_nonzero_eigenvalues}{the number of nonzero eigenvalues. This argument
is only used when \code{estimation="low"} and either \code{kinship} or
\code{K} is provided. This argument will limit the function to finding the
\code{n_nonzero_eigenvalues} largest eigenvalues.}

\item{n_zero_eigenvalues}{the number of desired or specifed zero eigenvalues.
This is only needed when \code{estimation="lowrank"}, and is calculated
internally by the \code{\link{ggmix}} function. It is equal to the number
of observations minus \code{n_nonzero_eigenvalues}}
}
\value{
A ggmix object, of the class that corresponds to the estimation
  method. These objects are lists that contain the data necessary for
  computation.
}
\description{
\code{new_fullrank_kinship}, \code{new_fullrank_K}, \code{new_fullrank_UD},
\code{new_lowrank_kinship}, \code{new_lowrank_K} and \code{new_lowrank_UD}
create the ggmix objects from the provided data that are necessary to fit the
penalized linear mixed model according to the user's parameters.
}
\seealso{
\code{\link{ggmix}}
}
