% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gic.ggmix}
\alias{gic.ggmix}
\title{Generalised Information Criterion}
\usage{
gic.ggmix(x, y, d, u, an = log(log(n)) * log(p), lambda = NULL, ...)
}
\arguments{
\item{an}{numeric, the penalty per parameter to be used; the default is an =
log(log(n))*log(p) where n is the number of subjects and p is the number of
parameters}

\item{...}{other arguments that can be passed to ggmix}
}
\description{
Calculates the generalised information criterion for each value of the tuning
parameter lambda
}
\details{
the generalised information criterion used for gaussian response is
  given by \deqn{-2 * loglikelihood(\hat{\Theta}) + an * df} where
  df is the number of non-zero estimated parameters
}
\references{
Fan Y, Tang CY. Tuning parameter selection in high dimensional
  penalized likelihood. Journal of the Royal Statistical Society: Series B
  (Statistical Methodology). 2013 Jun 1;75(3):531-52.

  Nishii R. Asymptotic properties of criteria for selection of variables in
  multiple regression. The Annals of Statistics. 1984;12(2):758-65.
}
