% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gic.penfam}
\alias{gic.penfam}
\title{Generalised Information Criterion}
\usage{
gic.penfam(x, y, d, u, an = log(log(n)) * log(p), lambda = NULL, ...)
}
\arguments{
\item{x}{input matrix, of dimension n x p; where n is the number of
observations and p are the number of predictors.}

\item{y}{response variable. must be a quantitative variable}

\item{d}{non-zero eigenvalues}

\item{u}{left singular vectors corresponding to the non-zero eigenvalues}

\item{an}{numeric, the penalty per parameter to be used; the default is an =
log(log(n))*log(p) where n is the number of subjects and p is the number of
parameters}

\item{lambda}{A user supplied lambda sequence (this is the tuning parameter).
Typical usage is to have the program compute its own lambda sequence based
on nlambda and lambda.min.ratio. Supplying a value of lambda overrides
this. WARNING: use with care. Do not supply a single value for lambda (for
predictions after CV use predict() instead). Supply instead a decreasing
sequence of lambda values. glmnet relies on its warms starts for speed, and
its often faster to fit a whole path than compute a single fit.}

\item{...}{other arguments that can be passed to penfam}
}
\description{
Calculates the generalised information criterion for each value of the tuning
parameter lambda
}
\details{
the generalised information criterion used for gaussian response is
  given by \deqn{-2 * loglikelihood(\hat{\Theta}) + an * df} where
  df is the number of non-zero estimated parameters
}
\references{
Fan Y, Tang CY. Tuning parameter selection in high dimensional
  penalized likelihood. Journal of the Royal Statistical Society: Series B
  (Statistical Methodology). 2013 Jun 1;75(3):531-52.

  Nishii R. Asymptotic properties of criteria for selection of variables in
  multiple regression. The Annals of Statistics. 1984;12(2):758-65.
}
