% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gic.R
\name{gic}
\alias{gic}
\alias{gic.default}
\alias{gic.ggmix_fit}
\title{Generalised Information Criterion}
\usage{
gic(ggmix_fit, ...)

\method{gic}{default}(ggmix_fit, ...)

\method{gic}{ggmix_fit}(ggmix_fit, ..., an = log(log(n)) * log(p))
}
\arguments{
\item{ggmix_fit}{An object of class \code{ggmix_fit} which is outputted by
the \code{\link{ggmix}} function}

\item{...}{other parameters. currently ignored.}

\item{an}{numeric, the penalty per parameter to be used; the default is an =
log(log(n))*log(p) where n is the number of subjects and p is the number of
parameters}
}
\value{
an object with S3 class \code{"ggmix_gic"}, \code{"ggmix_fit"},
  \code{"*"} and \code{"**"} where \code{"*"} is "lasso" or "gglasso" and
  \code{"**"} is fullrank or lowrank. Results are provided for converged
  values of lambda only. \describe{\item{ggmix_fit}{the ggmix_fit
  object}\item{lambda}{the sequence of converged tuning parameters}
  \item{nzero}{the number of non-zero estimated coefficients including the 2
  variance parameters which are not penalized and therefore always
  included}\item{gic}{gic value. a numeric vector with length equal to
  \code{length(lambda)}} \item{lambda.min.name}{a character corresponding to
  the name of the tuning parameter lambda which minimizes the
  gic}\item{lambda.min}{the value of lambda which minimizes the gic}}
}
\description{
Calculates the generalised information criterion for each value
  of the tuning parameter lambda
}
\details{
the generalised information criterion used for gaussian response is
  given by \deqn{-2 * loglikelihood(\hat{\Theta}) + an * df} where df is the
  number of non-zero estimated parameters, including variance components
}
\references{
Fan Y, Tang CY. Tuning parameter selection in high dimensional
  penalized likelihood. Journal of the Royal Statistical Society: Series B
  (Statistical Methodology). 2013 Jun 1;75(3):531-52.

  Nishii R. Asymptotic properties of criteria for selection of variables in
  multiple regression. The Annals of Statistics. 1984;12(2):758-65.
}
\seealso{
\code{\link{ggmix}}
}
