% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{grr_eta}
\alias{grr_eta}
\title{Gradient of eta parameter}
\usage{
grr_eta(eta, sigma2, beta, eigenvalues, x, y, nt)
}
\arguments{
\item{x}{should be U^T X, where U is the matrix of eigenvectors and X
contains the first column of ones for the intercept. Used for gradient of eta
(currently being passed to optim)}
}
\description{
Gradient of eta parameter
}
