% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{kkt_check}
\alias{kkt_check}
\title{Check of KKT for a given value of Lambda}
\usage{
kkt_check(eta, sigma2, beta, eigenvalues, x, y, nt, lambda,
  tol.kkt = 0.000000001)
}
\arguments{
\item{beta}{should include intercept as well. A 1 column matrix of dimension
(p+1) x 1.}

\item{x}{should be U^T X, where U is the matrix of eigenvectors and X
contains the first column of ones for the intercept. x should be a mtrix of
dimension n x (p+1)}

\item{tol.kkt}{Tolerance for determining if an entry of the subgradient is
zero}
}
\description{
Check of KKT for a given value of Lambda
}
