% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmlasso.R
\name{lmmlasso}
\alias{lmmlasso}
\title{Estimation of Linear Mixed Model with Lasso Penalty}
\usage{
lmmlasso(ggmix_object, ...)
}
\arguments{
\item{ggmix_object}{A ggmix_object object of class \code{lowrank} or
\code{fullrank}}

\item{...}{Extra parameters. Currently ignored.}

\item{n_design}{total number of observations}

\item{p_design}{number of variables in the design matrix, excluding the
intercept column}
}
\value{
A object of class \code{ggmix}
}
\description{
\code{lmmlasso} estimates the linear mixed model with lasso
  penalty
}
\seealso{
\code{\link{ggmix}}
}
