% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{lowrank}
\alias{lowrank}
\title{Fit a linear mixed model with lasso or elasticnet regularization}
\usage{
lowrank(x, y, d, u, lambda = NULL, lambda_min_ratio = ifelse(n < p, 0.01,
  1e-04), nlambda = 100, eta_init = 0.5, maxit = 100, fdev = 1e-05,
  standardize = FALSE, alpha = 1, thresh_glmnet = 1e-08,
  epsilon = 1e-04)
}
\description{
By the time the user gets to this point in the function they
  should have already computed the singular values and the eigenvectors. this
  is to allow flexibility with this software, so that if for example its
  easier to those things elsewhere they can, else they will be computed in a
  wrapper function
}
\seealso{
\code{\link[glmnet]{glmnet}}
}
