% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{penfam}
\alias{penfam}
\title{Fit a linear mixed model with lasso or elasticnet regularization}
\usage{
penfam(x, y, phi, lambda = NULL, lambda_min_ratio = ifelse(n < p, 0.01,
  0.001), nlambda = 100, eta_init = 0.5, maxit = 100, fdev = 1e-04,
  alpha = 1, thresh_glmnet = 1e-10, epsilon = 1e-05, an = log(log(n)) *
  log(n), tol.kkt = 1e-09)
}
\arguments{
\item{an}{is the constant used in the BIC calculation. The default choice is
from Wang et al. (2009)}
}
\description{
this is an older version of the code. currently not being used.
use \link{\code{lowrank}} instead
}
\references{
Wang, H., Li, B. and Leng, C. (2009) Shrinkage tuning parameter
  selection with a diverging number of parameters.J. R. Statist. Soc. B, 71,
  671–683.
}
