% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.gic.penfam}
\alias{plot.gic.penfam}
\title{Plot the Generalised Information Criteria curve produced by gic.penfam}
\usage{
plot.gic.penfam(x, sign.lambda = 1, ...)
}
\arguments{
\item{x}{fitted "gic.penfam" object}

\item{sign.lambda}{Either plot against log(lambda) (default) or its negative
if sign.lambda=-1}

\item{...}{Other graphical parameters to plot}
}
\description{
Plots the Generalised Information Criteria curve, as a function
  of the lambda values used
}
\details{
A plot is produced, and nothing is returned.
}
\seealso{
\code{\link{penfam}} and \code{\link{gic.penfam}}
}
