% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.ggmix}
\alias{coef.ggmix}
\alias{coef.gic.ggmix}
\alias{predict.ggmix}
\title{Get coefficients from a "ggmix" object}
\usage{
\method{coef}{ggmix}(object, s = NULL, which = NULL, ...)

\method{coef}{gic.ggmix}(object, s = "lambda.min", ...)

\method{predict}{ggmix}(object, new.x, new.u, new.d, s = NULL,
  type = c("link", "response", "coefficients", "ranef", "nonzero", "beta",
  "eta", "sigma2"))
}
\arguments{
\item{object}{object of class ggmix}

\item{s}{lambda value}

\item{which}{a character of either "beta", "eta", or "sigma2" if the user wants only th}

\item{...}{additional arguments to pass to predict function}

\item{type}{Type of prediction required. Type "link" gives the fitted values
given by \deqn{X\beta + b} where b is a subject-specific random effect. For
"gaussian" type "response" is equivalent to type "link". Type
"coefficients" computes the coefficients at the requested values for s.
Type "nonzero" returns a list of the indices of the nonzero coefficients
for each value of s.}

\item{s}{index of tuning parameter. Must be a character and an element of
"s1","s2",...."s100", where "s100" is the index of the last pair of tuning
parameters. Default is \code{NULL}}

\item{object}{Fitted ggmix model object}
}
\description{
Get coefficients from a "ggmix" object

this function only works for tuning parameter values defined by
  the shim_multiple_faster function. The interpolation feature is not working
  yet
}
