% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.ggmix_fit}
\alias{predict.ggmix_fit}
\title{Make predictions from a ggmix object}
\usage{
\method{predict}{ggmix_fit}(object, new.x, new.u, new.d, s = NULL,
  type = c("link", "response", "coefficients", "ranef", "nonzero", "beta",
  "eta", "sigma2"))
}
\arguments{
\item{object}{Fitted ggmix model object}

\item{s}{index of tuning parameter. Must be a character and an element of
"s1","s2",...."s100", where "s100" is the index of the last pair of tuning
parameters. Default is \code{NULL}}

\item{type}{Type of prediction required. Type "link" gives the fitted values
given by \deqn{X\beta + b} where b is a subject-specific random effect. For
"gaussian" type "response" is equivalent to type "link". Type
"coefficients" computes the coefficients at the requested values for s.
Type "nonzero" returns a list of the indices of the nonzero coefficients
for each value of s.}
}
\description{
this function only works for tuning parameter values defined by
  the shim_multiple_faster function. The interpolation feature is not working
  yet
}
