% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.gic.ggmix}
\alias{predict.gic.ggmix}
\title{Make predictions from a "gic.ggmix" object.}
\usage{
\method{predict}{gic.ggmix}(object, newx, s = c("lambda.1se", "lambda.min"),
  ...)
}
\arguments{
\item{object}{a fitted \code{\link{gic.ggmix}} object}

\item{newx}{matrix of new values for x at which predictions are to be made}

\item{s}{value(s) of the penalty parameter lambda at which predictions are
required. Default is the value s="lambda.min" stored on the
\code{\link{gic.ggmix}} object. If s is numeric, it is taken as the
value(s) of lambda to be used.}

\item{...}{not used. Other arguments to predict}
}
\description{
This function makes predictions from a ggmix model with GIC,
  using the stored "ggmix.fit" object, and the optimal value chosen for lambda
  based on the minimum GIC.
}
\details{
This function makes it easier to use the results of cross-validation
  to make a prediction
}
