% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.penfam}
\alias{coef.penfam}
\alias{predict.penfam}
\title{Get coefficients from a "penfam" object}
\usage{
\method{coef}{penfam}(object, s = NULL, ...)

\method{predict}{penfam}(object, newx, s = NULL, type = c("link",
  "response", "coefficients", "nonzero"))
}
\arguments{
\item{object}{Fitted shim model object}

\item{s}{index of tuning parameter. Must be a character and an element of
"s1","s2",...."s100", where "s100" is the index of the last pair of tuning
parameters. Default is \code{NULL}}

\item{type}{Type of prediction required. Type "link" gives the fitted values.
Type "response" for "gaussian" type "response" is equivalent to type
"link". Type "coefficients" computes the coefficients at the requested
values for s. Type "nonzero" returns a list of the indices of the nonzero
coefficients for each value of s.}
}
\description{
Get coefficients from a "penfam" object

this function only works for tuning parameter values defined by
  the shim_multiple_faster function. The interpolation feature is not working
  yet
}
