% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.ggmix_fit}
\alias{print.ggmix_fit}
\alias{print.ggmix_gic}
\title{Print Method for Objects of Class \code{ggmix_fit}}
\usage{
\method{print}{ggmix_fit}(x, ..., digits = max(3, getOption("digits") - 3))

\method{print}{ggmix_gic}(x, ..., digits = max(3, getOption("digits") - 3))
}
\arguments{
\item{x}{an object of class objects of class \code{ggmix_fit}}

\item{...}{other arguments passed to \code{print}}

\item{digits}{significant digits in printout. Default: \code{max(3,
getOption("digits") - 3)}}
}
\value{
The call that produced the object \code{x} is printed, followed by a
  three-column matrix with columns \code{Df}, \code{\%Dev}, and and
  \code{Lambda}. The \code{Df} columns correspond to the number of nonzero
  coefficients including variance components. \code{\%dev} is the percent
  deviance explained (relative to the null deviance). \code{Lambda} is the
  sequence of converged tuning parameters.
}
\description{
print method for objects of class \code{ggmix_fit}
}
\seealso{
\code{\link{ggmix}}
}
