% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.penfam}
\alias{print.penfam}
\title{Print Method for penfam function}
\usage{
\method{print}{penfam}(x, digits = max(3, getOption("digits") - 3), ...)
}
\description{
print method for penfam function
}
