% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ranef}
\alias{ranef}
\alias{random.effects}
\alias{random.effects.default}
\alias{ranef.default}
\title{Extract Random Effects}
\usage{
ranef(object, ...)

random.effects(object, ...)

\method{random.effects}{default}(object)

\method{ranef}{default}(object)
}
\arguments{
\item{object}{any fitted model object from which random effects estimates can be
extracted. Currently supports "gic.ggmix" object}

\item{...}{some methods for this generic function require additional
arguments.}
}
\value{
a numeric vector of length n of subject-specific random effects
}
\description{
Generic function for extracting the random effects. This is the same generic
(and same name) defined in the nlme package.
}
\seealso{
\code{\link{gic.ggmix}}
}
