% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{sigma2}
\alias{sigma2}
\title{Closed form solution for sigma^2, given beta and eta}
\format{An object of class \code{numeric} of length 1.}
\usage{
sigma2
}
\description{
Closed form solution for sigma^2, given beta and eta
}
\keyword{datasets}
