% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lambda_sequence}
\alias{lambda_sequence}
\title{Calculate Sequence of Tuning Parameters For}
\usage{
lambda_sequence(x, y, eigenvalues, weights = NULL, lambda_min_ratio,
  epsilon = 1e-14, tol.kkt = 1e-09, eta_init = 0.5, nlambda = 100,
  scale_x = F, center_y = F)
}
\arguments{
\item{weights}{Separate penalty factors can be applied to each coefficient.
This is a number that multiplies lambda to allow differential shrinkage,
and can be used to apply adaptive LASSO. Can be 0 for some variables, which
implies no shrinkage, and that variable is always included in the model.
Default is 1 for all variables (and implicitly infinity for variables
listed in exclude). Note: the penalty factors are internally rescaled to
sum to nvars, and the lambda sequence will reflect this change.}

\item{nlambda}{the number of lambda values - default is 100.}

\item{scale_x}{should the columns of x be scaled - default is FALSE}

\item{center_y}{should y be mean centered - default is FALSE}

\item{lambda.factor}{The factor for getting the minimal lambda in lambda
sequence, where \code{min(lambda) = lambda.factor * max(lambda).
max(lambda)} is the smallest value of lambda for which all coefficients are
zero. The default depends on the relationship between \code{N} (the number
of rows in the matrix of predictors) and \code{p} (the number of
predictors). If \code{N > p}, the default is \code{1e-6}, close to zero. If
\code{N < p}, the default is \code{0.01}. A very small value of
lambda.factor will lead to a saturated fit.}
}
\value{
numeric vector of length \code{q}
}
\description{
Function to calculate the sequence of tuning parameters based on
  the design matrix \code{x} and the response variable {y}. This is used in
  the \code{\link{shim_once}} function to calculate the tuning parameters
  applied to the main effects
}
\details{
The maximum lambda is calculated using the following inequality:
  \deqn{(N*w_j)^-1 | \sum x_ij y_i | \le \lambda_max}

  The minimum lambda is given by lambda.factor*lambda_max. The sequence of
  nlambda values are decreasing from lambda_max to lambda_min on the log
  scale.

  The penalty factors are internally rescaled to sum to the number of
  predictor variables in glmnet. Therefore, to get the correct sequence of
  lambdas when there are weights, this function first rescales the weights
  and then calclated the sequence of lambdas.

  This formula is taken from section 2.5 of the \code{glmnet} paper in the
  Journal of Statistical Software (see references for details)
}
