% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmlasso.R
\name{lmmlasso}
\alias{lmmlasso}
\alias{lmmlasso.default}
\alias{lmmlasso.fullrank}
\title{Estimation of Linear Mixed Model with Lasso Penalty}
\usage{
lmmlasso(ggmix_object, ...)

\method{lmmlasso}{default}(ggmix_object, ...)

\method{lmmlasso}{fullrank}(ggmix_object, penalty.factor, lambda,
  lambda_min_ratio, nlambda, n_design, p_design, eta_init, maxit, fdev,
  standardize, alpha, thresh_glmnet, epsilon)
}
\arguments{
\item{ggmix_object}{A ggmix_object object of class \code{lowrank} or
\code{fullrank}}

\item{...}{Extra parameters.}

\item{shrink}{Should we use a shrinkage estimate of the residual variance?}

\item{index}{If \code{pcevObj} is of class \code{PcevBlock}, \code{index} is
a vector describing the block to which individual response variables
correspond.}
}
\value{
A list containing the variance components, the first PCEV, the
  eigenvalues of \eqn{V_R^{-1}V_M} and the estimate of the shrinkage
  parameter \eqn{\rho}
}
\description{
\code{lmmlasso} estimates the linear mixed model with lasso
  penalty
}
\seealso{
\code{\link{ggmix}}
}
