% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ggmix_gic}
\alias{plot.ggmix_gic}
\alias{plotGIC}
\title{Plot the Generalised Information Criteria curve produced by \code{gic}}
\usage{
\method{plot}{ggmix_gic}(x, sign.lambda = 1, ...)

plotGIC(x, sign.lambda, lambda.min, ...)
}
\arguments{
\item{x}{fitted linear mixed model object of class \code{ggmix_gic} from the
\code{\link{gic}} function}

\item{sign.lambda}{Either plot against log(lambda) (default) or its negative
if sign.lambda=-1}

\item{...}{Other graphical parameters to plot}

\item{lambda.min}{the value of lambda which minimizes the gic}
}
\description{
Plots the Generalised Information Criteria curve, as a function
  of the lambda values used
}
\details{
A plot is produced, and nothing is returned.
}
\seealso{
\code{\link{ggmix}} and \code{\link{gic}}
}
