% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.ggmix}
\alias{coef.ggmix}
\alias{coef.gic.ggmix}
\title{Get coefficients from a "ggmix" object}
\usage{
\method{coef}{ggmix}(object, s = NULL, which = NULL, ...)

\method{coef}{gic.ggmix}(object, s = "lambda.min", ...)
}
\arguments{
\item{object}{object of class ggmix}

\item{s}{lambda value}

\item{which}{a character of either "beta", "eta", or "sigma2" if the user
wants only th}

\item{...}{additional arguments to pass to predict function}
}
\description{
Get coefficients from a "ggmix" object
}
