% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\name{ranef}
\alias{ranef}
\alias{random.effects}
\alias{random.effects.default}
\alias{ranef.default}
\alias{ranef.gic.ggmix}
\alias{ranef.ggmix}
\alias{random.effects.gic.ggmix}
\title{Extract Random Effects}
\usage{
ranef(object, ...)

random.effects(object, ...)

\method{random.effects}{default}(object)

\method{ranef}{default}(object)

\method{ranef}{gic.ggmix}(object, s = "lambda.min", ...)

\method{ranef}{ggmix}(object, new.x, new.u, new.d, s = NULL,
  type = c("fitted", "predicted"))

\method{random.effects}{gic.ggmix}(object, s = "lambda.min")
}
\arguments{
\item{object}{any fitted model object from which random effects estimates can be
extracted. Currently supports "gic.ggmix" object}

\item{...}{some methods for this generic function require additional
arguments.}

\item{s}{lamda at which to predict the random effects. current option is only
"lambda.min"}

\item{s}{index of tuning parameter. Must be a character and an element of
"s1","s2",...."s100", where "s100" is the index of the last pair of tuning
parameters. Default is \code{NULL}}
}
\value{
a numeric vector of length n of subject-specific random effects
}
\description{
Generic function for extracting the random effects. This is the same generic
(and same name) defined in the nlme package.
}
\details{
For objects of class "gic.ggmix", this function returns the
  subject-specific random effect value for the model which minimizes the GIC
  using the maximum a posteriori principle

For objects of class "ggmix", this function returns the
  subject-specific random effect value for the model which minimizes the GIC
  using the maximum a posteriori principle
}
\seealso{
\code{\link{gic.ggmix}}
}
