% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\name{ranef}
\alias{ranef}
\alias{random.effects}
\alias{random.effects.default}
\alias{ranef.default}
\alias{ranef.gic.penfam}
\alias{random.effects.gic.penfam}
\title{Extract Random Effects}
\usage{
ranef(object, ...)

random.effects(object, ...)

\method{random.effects}{default}(object)

\method{ranef}{default}(object)

\method{ranef}{gic.penfam}(object, s = "lambda.min")

\method{random.effects}{gic.penfam}(object, s = "lambda.min")
}
\arguments{
\item{object}{any fitted model object from which random effects estimates can be
extracted. Currently supports "gic.penfam" object}

\item{...}{some methods for this generic function require additional
arguments.}

\item{s}{lamda at which to predict the random effects. current option is only
"lambda.min"}
}
\value{
a numeric vector of length n of subject-specific random effects
}
\description{
Generic function for extracting the random effects. This is the same generic
(and same name) defined in the nlme package.
}
\details{
For objects of class "gic.penfam", this function returns the
  subject-specific random effect value for the model which minimizes the GIC
  using the maximum a posteriori principle
}
\seealso{
\code{\link{gic.penfam}}
}
